+++ src/test/java/org/mule/modules/sqs/automation/testcases/GetQueueUrlTestCases.java
@@ -9,6 +9,7 @@ package org.mule.modules.sqs.automation.testcases;
 
 import com.amazonaws.services.sqs.model.CreateQueueResult;
 import com.amazonaws.services.sqs.model.GetQueueUrlResult;
+import org.apache.commons.lang.StringUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -24,6 +25,7 @@ import static org.junit.Assert.*;
 public class GetQueueUrlTestCases extends SQSFunctionalTestParent {
 
     private String queueUrl;
+    final String TEST_QUEUE_NAME = ConnectorTestUtils.generateRandomShortString();
 
     @Before
     public void setup() throws Exception {
@@ -52,7 +54,9 @@ public class GetQueueUrlTestCases extends SQSFunctionalTestParent {
 
     @After
     public void tearDown() throws Exception {
-        getConnector().deleteQueue(queueUrl);
+        if (StringUtils.isNotBlank(queueUrl)) {
+            getConnector().deleteQueue(queueUrl);
+        }
     }
 
 }
