+++ src/test/java/org/mule/modules/sqs/automation/testcases/PurgeQueueTestCases.java
@@ -8,6 +8,7 @@ package org.mule.modules.sqs.automation.testcases;
 
 import com.amazonaws.services.sqs.model.CreateQueueResult;
 import com.amazonaws.services.sqs.model.SendMessageResult;
+import org.apache.commons.lang.StringUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -25,6 +26,7 @@ import static org.junit.Assert.fail;
 public class PurgeQueueTestCases extends SQSFunctionalTestParent {
 
     String queueUrl;
+    final String TEST_QUEUE_NAME = ConnectorTestUtils.generateRandomShortString();
 
     @Before
     public void setUp() throws Exception {
@@ -51,6 +53,8 @@ public class PurgeQueueTestCases extends SQSFunctionalTestParent {
 
     @After
     public void tearDown() throws Exception {
-        getConnector().deleteQueue(queueUrl);
+        if (StringUtils.isNotBlank(queueUrl)) {
+            getConnector().deleteQueue(queueUrl);
+        }
     }
 }
