+++ src/test/java/org/mule/modules/sqs/automation/testcases/SetQueueAttributesTestCases.java
@@ -9,6 +9,7 @@ package org.mule.modules.sqs.automation.testcases;
 
 import com.amazonaws.services.sqs.model.CreateQueueResult;
 import com.amazonaws.services.sqs.model.GetQueueUrlResult;
+import org.apache.commons.lang.StringUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -29,6 +30,7 @@ import static org.junit.Assert.fail;
 public class SetQueueAttributesTestCases extends SQSFunctionalTestParent {
 
     String queueUrl;
+    final String TEST_QUEUE_NAME = ConnectorTestUtils.generateRandomShortString();
 
     @Before
     public void setUp() throws Exception {
@@ -60,6 +62,8 @@ public class SetQueueAttributesTestCases extends SQSFunctionalTestParent {
 
     @After
     public void tearDown() throws Exception {
-        getConnector().deleteQueue(queueUrl);
+        if (StringUtils.isNotBlank(queueUrl)) {
+            getConnector().deleteQueue(queueUrl);
+        }
     }
 }
